Fix an issue when restoring the main activity from a corrupted state.
When APDE crashed in an activity deep into the hierarchy - for example, one of the Settings screens - this used to also crash the main activity, making for two crashes, because the saved state became corrupted. This second crash did away with the stack trace of the first, so it became difficult to debug these deep-hierarchy problems. With any luck, this problem is now fixed and we can start actually solving real issues.