Hide normal error messages on first startup.
When APDE is opened for the first time, it can't load the temporary files.  Now, we're suppressing these error messages to keep the clutter away from new users...

Note: These error messages actually appearing are new because we changed the launch order to set up the console before the sketch is loaded. So these error messages didn't appear before, but it's much cleaner this way.