Further work toward implementing Material Design.
Now using ViewPager to switch between tabs, which is a huge change from the hacky, unsupported, and now-impossible method from before. Various work in all parts of the application, currently not fully functioning, but has to be committed. Also, switch to using some of Google's Material icons.

FileMeta is now SketchFile, to better represent what it actually is. There are now fragments associated withe every SketchFile for the ViewPager, just to make things more confusing.