Materialize everything!
The entire app is now fully Material Design-ified and functional, with the notable exception of saving temporary sketches. A new temporary sketch system will be implemented shortly.

In the future, the minSdk will have to be changed to API level 15 (Android 4.0.3), which is what the newer versions of the desktop Android Mode currently support. I haven't actually made this change yet, but I have started dropping support for 2.3.3, for example in the Settings screen.

The changes included in this commit are numerous and I have lost track of many of them. They include using all of Google's support libraries, including the very buggy preference library (for which I am using a third party fix) for the new-and-improved Settings screen. I have changed to using the support library's TabLayout for the tab bar, which is much cleaner than what I had before. I have re-done a lot of the layouts so that they look correct with the new Material widgets and icons. There are many other changes and bugfixes that I can't think of right now.