Implement new and improved temporary sketch model.
There is now another top-level sketch location for temporary sketches alongside the sketchbook, examples, etc. Temporary sketches no longer use the default name "sketch" and instead are named similarly to the desktop PDE as "sketch_" followed by the sketch's date of creation. A new temporary sketch is automatically created whenever "New Sketch" is selected or a sketch is deleted. Temporary sketches are saved in a special internal storage directory and can thus be run without the clunkly old system that involved saving and renaming the sketch. Temporary sketches can be moved to the sketchbook.