Change minSdk to API level 15.
This means dropping support for 2.3.3, which is currently 1.14% of devices on which APDE is installed. Supporting 2.3.3 is a pain and requires a great deal of code duplication that I didn't want to deal with when upgrading to Material Design, so the repository is currently broken for 2.3.3 anyway. The desktop Android mode has already dropped support for 2.3.3, switching to a minimum of API level 15 as well.

As for API levels 11-14, these represent practically none of the userbase for both APDE and Android as a whole.