Warn users about screen overlay apps.
When a screen overlay app is enabled, the Android system prevents the user from pressing the "Install" button in the package installer.

When the user presses the run button, check to see if a screen overlay is enabled, using the same logic as the Android system, and create a dialog informing the user if this is the case.