Change to new examples repository.
The old examples repo wasn't as forward-thinking as I thought it was. The problem was that when we updated the examples in such a way that they were no longer backwards-compatible (like upgrading to the Android Mode 3.0.1 examples), it wasn't possible to prevent the older versions of the app from also getting the newer examples that wouldn't work for those versions.

Therefore, I have created a new examples repository for v0.4.0+, which will have a separate branch each time that the examples are updated in a non-backwards-compatible way. The code that checks for examples updates has been changed accordingly.