Implement Sandbox mode.
Sandbox mode is a Processing.js-based environment for quickly running APDE sketches. It exists in parallel to the standard Android build method. There is a new button in the toolbar to open sandbox mode. On phones this launches a separate activity running the Processing.js sketch full screen and on tablets in landscape mode this opens a separate pane (SandboxFragment) that contains a smaller version of the Processing.js sketch. The tablet pane also has a toolbar at the top with options to reload the sketch (restart it with updated code), go fullscreen (like the phone version), and close out of the Sandbox.

The version of Processing.js used is a custom build with specific changes designed to work with APDE; see https://github.com/Calsign/processing-js, specifically the "apde" branch for the changes made.

Eventually, the goal is to build real-time code injection into the Sandbox to allow for changes to the code to be seen in real-time, but this will take a considerable amount of work to pull off.

There is currently no side-by-side view for phones; this is more difficult to implement due to the small screen size. One potential solution is to show the Sandbox pane over the console.

As part of the introduction of the Sandbox, the run and stop buttons have been combined (which makes more sense anyway). The run button becomes the stop button when the sketch is building and changes back after building stops (these changes have morph animations for Android 5.0+). This consolidation of screen real estate allows the sandbox icon to be shown in the toolbar on most screens.

The sandbox icon is one that I designed. It looks like a pail/bucket full of sand. I tried to follow the Material Design icon guidelines.

Numerous other random changes have been made as well. There are still a handful of bugs concerning the Sandbox.