Implement the framework for the learning curriculum.
There is now a "Learning" screen accessible from the action overflow menu. The learning curriculum combines textual tutorials with the Sandbox for seeing changes made to code in real time. The tutorials themselves are downloaded from a new learning repository that is managed the same way as the examples repository:

https://github.com/Calsign/APDE-Learning-Repo/

Note that as of this commit, only four of the tutorials have been written; more will be added over time and downloaded as examples/learning curriculum updates.

The tutorials are structured in a tree. Each tutorial (except for the first) has prerequisite tutorials, but the user is not limited to taking one path. There are many different ways to work through the tutorials. The tree is displayed on the initial learning screen, the curriculum overview, with a SurfaceView and canvas and arrows representing the prerequisites.

The tutorials are presented in a separate fragment within the learning activity. Each tutorial is loaded from an XML file that is part of the learning repository. This file can contain pseudo-Markdown text, images, code snippets, code display areas, and lists of examples. ("Quizzes", a method to test user understanding of the tutorial material, are planned but not currently implemented.) These components are grouped into "segments" that are displayed on separate pages which can be navigated by the user. The text is parsed with Markdown-inspired syntax, allowing for text formatting and inline code snippets. The code areas are displayed using the CodeEditText, so they are automatically syntax highlighed. The code areas are optionally runnable, which will open a Sandbox fragment. Additionally, the code areas are optionally "pickable", which means that certain values (integers and colors, at the moment) may be selected and a slider dragged to change the value and see the output in real time in the connected code display area (embedded Sandbox fragment). This is the pattern presented by Bret Victor in his "Inventing on Principle" talk that was also implemented by the Khan Academy computer programming tutorials. The lists of examples provide links to examples that are recommended upon completion of the tutorial. For now (until quizzes are implemented), navigation to the last page of a tutorial and pressing the "Finish" button is considered to be completion of that tutorial.