Fix several build issues.
Potential fix for the libpng security vulnerability: switch to new AAPT binaries. I won't know if this fix worked until I try uploading the APK to Google Play.

Fix two build issues affecting all Android 7.0 Nougat devices - DX Merger NoSuchMethodError and a package installer permissions problem (the latter was not discovered until after fixing the former).

The NoSuchMethodError proved a tough nut to crack. The problem was that core-libart (the core Android library) contains the Dex library but not DX Merger. The dx.jar library included in the libs folder contains both Dex and DX Merger as is necessary, but both libraries (in the project and in core-libart) used the same package name. This caused the ClassLoader to always load the core-libart version of Dex and the dx.jar version of DX Merger - two different versions of the same library trying to work together. With API 24, the method signature for com.android.dex.TableOfContents.writeHeader() changed, leading to a NoSuchMethodError when the older dx.jar DX Merger tried to access the newer core-libart method. This problem was solved by changing the package name of the included dx.jar library with the Jar Jar Links JAR re-packager tool so that the ClassLoader always loads the dx.jar version of both Dex and DX Merger. In accordance with the tool's name, the new package name starts with "com.androidjarjar".

The package manager permissions problem was due to the removal of MODE_WORLD_READABLE by way of causing a SecurityException on Nougat. In its place I have implemented a FileProvider to do things the "right way", but this seems to cause problems below API level 24 so the old method is left in place for pre-Nougat devices. The "right way" also requires the REQUEST_INSTALL_PACKAGES permission on API 23+. This permission is therefore only requested on API 23+.

Also target API 25, upgrade to the latest support library, and remove the MIPS AAPT binary.