Workaround for #24.
The combination of onSaveInstanceState() and onRestoreInstanceState() was causing strange problems, including the occasional deletion of code under seemingly random (and difficult to reproduce) circumstances. This was too risky, so we are now completely avoiding having to do save the instance state by forcing APDE to re-load the sketch from disk every time that the instance state saving would be used otherwise, including after screen rotations and opening the recent apps menu.

Consequently the open file state (the selection, scroll position, and undo/redo history) is no longer saved and restored by the Android system, so the sketch is reset to default state (losing all of that data). The mechanism for storing the last-opened sketch (saving a temporary file on the internal storage) has been overhauled to include not just the sketch title and location, but also the currently selected tab and the selection and scroll positions for all open tabs. This means that the last-open state of the sketch is also restored at times when APDE previously reset, including when the app is closed with the back button.

This solution is not yet complete, however, as the undo/redo history is lost when the sketch is re-loaded, including after the screen is rotated. This is unacceptable. Thus a new solution for storing the undo/redo history long-term will be implemented, meaning that the revision history will be preserved all the way back to sketch creation.