Preserve undo/redo history with a temp JSON file.
The purpose of this change is to keep the undo/redo history when the user changes the screen orientation, APDE goes into the background, or the screen turns off. This is an issue because we are no longer using parcels to preserve this information because parcels are a bad day for everyone.

The undo/redo history is stored in JSON format as a temp file within APDE's internal memory. Each file within a sketch has its history stored as a separate JSON element within the file.

Note that the undo/redo history is still preserved when APDE is killed completely, which differs from the behavior of the parcelized version of the history preservation. The undo/redo history is only reset when a new sketch is loaded because APDE only stores the history for one sketch at a time. This side-effect is not currently viewed as undesirable.

A corner case that could potentially cause problems is the instance in which the user edits the sketch files with an external editor. In this case performing undo or redo after loading the sketch in APDE again would break things because the history is out of sync with the file. To prevent this from becoming an issue, APDE performs an MD5 hash of the sketch file when saving and compares it to a hash of the current text when loading to detect external changes and prevent the undo/redo history from being loaded if so.