Implement several new keybindings.
Added keybindings for accessing the sketchbook, examples, and recent sketches, opening settings and sketch properties, showing the sketch folder, opening the color selector, finding the selection in the reference, and opening the reference. Placeholders for some additional keybindings have been put in place, but these have not yet been implemented.

Also fixed a couple of quirks in find/replace and the drawer.