Support build to live wallpapers.
Target selector not yet implemented, so currently hardcoded to building only live wallpapers.

After installing the sketch, APDE immediately opens the live wallpaper preview screen with the option to set as a live wallpaper. The whole process is seamless.

Log broadcaster modified to not depend on having an activity.

Live wallpapers don't seem to like having a call to size(), so the preprocessor has been modified to force wallpapers to use fullScreen().

Created a tool called "Live Wallpapers" in the menu that opens the Android live wallpaper selector.