Remove size from setup in sketches, was causing problems for wallpapers.
Everything wasn't quite peachy last time. It turns out wallpapers don't like size calls if it's anything other than displayWidth and displayHeight. So the solution is to force fullScreen and remove the size call from setup. The size removal is a little hacky but it seems reliable so far.

Additionally, don't launch the wallpaper selector if the user didn't actually install the sketch. This change also seems to streamline the install process by skipping the user's needing to press "OK", going straight to the wallpaper preview. I don't know whether or not this applies to non-Samsung devices, but it is definitely nice.