Let user select component target (app, wallpaper, etc.) from toolbar.
There is a new button in the toolbar next to run. Stop has been relegated to the overflow menu (unless the screen has room for more than two buttons). The button shows the currently selected component target and opens a dropdown to pick the target when pressed. The icons representing the different targets are subject to change.

All four eventual targets are in the dropdown but only app and wallpaper are currently implemented - watchface and VR are currently placeholders.