Add VR support.
Requires the Processing VR and Google VR libraries. The GVR libs have been combined with the binaries removed (to reduce file size). The binaries are added separately, just like the GLSL shader files. These files are pretty big; there might be ways to make them smaller but they are worth the APK size increase.

VR sketches are forced to be fullscreen with the STEREO renderer. The Android mode docs say that this is required for all VR sketches, but we are making the preprocessor enforce this so that the user can run most examples right out of the box without having to mess with the renderer. VR sketches also ignore the screen orientation manifest setting.

All tested 3D examples work. Non-3D sketches run fine but produce no apparent display output - this is to be expected of course. Also tested with picorana's VR_audioscape demo.

Something may be funky with console output. Doesn't always show up after first run but seems to appear on subsequent runs. Need to investigate more thoroughly but not a major issue.