Support the MONO (magic window) VR renderer.
I had previously overlooked the existence of mono, forcing every VR sketch into stereo. Now the preprocessor preserves the renderer if it is a valid VR renderer (stereo or mono) and uses the default renderer otherwise.

The default renderer is configured by the user in a new setting under Build > VR. If the user does not change this setting, it is set to stereo. The rationale behind this is that not everyone has a VR headset. People that have a headset will likely opt to default to stereo (headset), but those without a headset may choose mono (magic window) instead.

For some reason mono sketches are really hard to quit. VR seems to take over the phone, meaning that the navigation drawer and other things disappear (if the phone has no hardware navigation buttons, like mine). The stereo renderer (perhaps the Google VR library itself?) offers an "X" button in the corner for quitting sketches, but mono has no such button. I have been able to quit mono sketches by long-pressing power to open the restart/shutdown menu and by triggering "OK, Google", but this is far from ideal. It is easy to imagine that there are devices for which neither of these is a good option and the user would be forced to power cycle the device to quit a mono sketch. This also appears to be the way that desktop Android mode works, so perhaps it is worth opening an issue in their repo.