Streamline sketch install/run process.
We got the wallpaper chooser to open immediately after installing the sketch (skipping the "open" button in the package installer), and it turns out the same thing works for regular apps. Now for both the app and VR targets, the sketch will launch without the user having to press "open".

If the user loads a different sketch while the sketch is building, then APDE will fail to start the sketch because it checks the current sketch rather than the sketch that was installed. This is a corner case that probably isn't worth correcting because the user would have to actively try to break things for this to happen.