Support assets (data folder) for watchface sketches.
We are loading the sketch files from the APK, so the sketch doesn't know where to find its assets (generally things in the data folder). So now we copy everything in assets into the wear companion's private internal storage directory. The only thing that the wear companion stores there is the sketch APK itself, so there shouldn't be any conflicts between wear companion and sketch files. All of the assets are deleted when a new sketch is loaded. We are also deleting any shared preferences created by the sketch to create a blank slate every time that a sketch is loaded.

One slight problem is that this is also where sketches save files to. So there could be an issue if the sketch tries saving a file with the same name as a file in the data folder. That sounds like it would be confusing (two files with the same name in different places) and a bad idea. Hopefully nobody actually tries to do this.

A better solution might be to modify PApplet (in processing/processing-android), but this should still work and it is much less maintenance.

Also removed the toast from the watchface loader. This was causing a crash (something about displaying UI from a service) and was just bad news in general. We don't need the toast because the watch face picker shows up right away anyway.