Warn the user if the watch is not connected when running watchfaces.
When the user runs a sketch with the watchface target, check to see if the wear companion is visible. If not, tell the user and stop the build. We check by seeing if the capability defined by the wear companion app is accessible from APDE. This applies when the wear companion isn't installed on the watch or when the devices are not connected through bluetooth. There still seem to be some cases where APDE does not detect a connection failure but the wear companion still fails to receive the sketch (ex. just disconnected the watch and the phone still thinks it is connected) - I think that these are corner cases that we shouldn't have to worry about.

It turns out that the watchface loader service is started by intent directly from APDE sending the sketch. In other words, we don't have to do anything to start it, it is entirely automatic. So I have removed the code that dealt with starting and stopping the service (no wonder it didn't seem to be working).

The wear companion activity has turned into a short message telling the user to run a sketch with the watchface target. It also displays the version of the wear companion. We could probably get rid of the activity entirely, but some users might want to have it to know that the wear companion is installed or something like that.

Also restructure some things.