Fix VR crash when sketch starts for some devices.
I could not reproduce this, but when targeting VR on Sara's (my mentor's) phone, the sketch crashed with a NoClassDefFoundError for the R.java for the Google VR library. Now AAPT builds R for the GVR library when VR is selected. It actually builds R for GVR instead of AppCompat, but this seems to be OK because AppCompat's R isn't used for VR sketches. (It is easier to implement it this way because Java arrays are fixed-length and I had difficulties getting AAPT to recognize the "--extra-packages" flag twice.)

Also add "processing.vr" to the list of ignorable imports when targeting VR. On the desktop, VR is packaged as a library, but for APDE we are not doing it this way because we want any 3D sketch to be runnable as a VR sketch.