Store manifest settings in properties instead of manifest, fixes #45.
We no longer store information directly in the AndroidManifest.xml file, instead we use sketch.properties, which contains all of the manifest options available in the sketch properties activity. Then when we build the sketch, we generate the manifest from the properties file. This is much much much cleaner than what we were doing before. For a more detailed explanation, see my comment on issue #45.

The hard part was upgrading from the manifest to the properties file. There were a lot of corner cases that needed checking, but I think I have accounted for everything. This could use more extensive testing.