Fix some problems with preview on Android 5.1 and 4.4.
First, the file permissions logic was wrong for API < 24. Then on Android 4.4 we need the support-wearable library (which is very annoying), so rebuilt the preview-build.zip file.

I am still experiencing an issue in a 4.4 emulator: the Dalvik VM is unable to JIT the loaded classes because its cache directory is full. I am hoping that this is just an emulator issue. Note that it took some hunting to actually find this error message, instead of just getting a ClassNotFoundException. I had to edit the sketch previewer and force a RuntimeException so that I could explore the full error logs.

I haven't tested emulators beyond 5.1 and 4.4, but I think that these two give us most of the necessary coverage because most of the version-specific problems show their faces here.