Add sketch previewer settings: change permissions, reinstall, uninstall.
The new settings are available in Settings > Build > Preview Mode.

The sketch preview builder has been turned into a task. The task system has been improved slightly.

The permission selector uses the same PermissionsActivity as sketch properties. The user is given the option to reinstall the sketch previewer when they close out of the permissions screen. The permissions that they selected are not saved if they choose to exit, however, because this would be confusing: when the user re-opens the permissions screen, should it show the saved ones or the installed ones? So the selected permissions aren't persisted, but the user can reinstall the sketch previewer when closing the permissions screen.