Get errors and warnings from ECJ and highlight the corresponding code.
We prevent ECJ from printing to the console and instead pipe the output into our own "problem" handler (problem is catch-all term for errors and warnings). We currently draw lines under all of the problems, just like in a full-blown IDE. This seems to be working fairly well, although there may still be bugs.

We do not yet compile in real-time. The problems are only updated when you press "run". This means that if you edit the code, the problem underlines get out of sync until you re-run. This is intended to change, of course, when we start running the compiler in the background so that the problems appear as you type, just like in a real IDE!

Problems that are detected by the preprocessor are currently handled differently. We should fix this so that the preprocessor feeds directly into this compiler problem system.

We also don't have a list of all problems yet. The idea is to do this like the desktop PDE and make another tab in the console area. This should not be too hard - we already have a list of all problems by tab - but it needs to be done.