Combine run and stop buttons.
The run and stop buttons are never both needed at the same time. Run now smoothly animates into the stop button when pressed, and the reverse happens when the build finishes (or fails). Below API level 21, the animations API is not as nice, so the image simply changes (not a smooth animation).

This implementation is taken directly from my prior commit (5dc19a6) to the learning branch for the sandbox mode. It required some minor tweaks but otherwise worked right out of the box. I didn't do any fancy git things to move the changes correctly because the original commit contained a lot of extra stuff that I didn't want.

It makes sense to make this change now that we have the target selection button next to the run button instead of the stop button.