Display compiler problems in problem overview screen.
The console area has been replaced by a view pager containing the console area and the problem overview screen (similar to the desktop PDE). You can switch between them by swiping or pressing the new "problems" button on the menu bar. The problems overview is a list of all of the problems reported by ECJ for the last build. Selecting a problem will jump to that problem in the code. Long-pressing a problem will copy the error message to the clipboard.

The next steps are integrating the preprocessor errors into the same problem pipeline and daemonizing the build system so that compilation occurs in real-time as the user types.

A lot of random problems with the placement and sizing of the code, console, and message areas have been fixed. There might still be some quirks but things should be less buggy in general.