Lots of work toward a new, JDT DOM-based preprocessor.
When working on incremental compilation, it became apparent that the existing preprocessor was inadequate for our needs. Errors picked up by ECJ are all reported at once, but the preprocessor was only capable of spitting out one error at a time. Not to mention the fact that the old preprocessor was ridiculously hacked together.

The new preprocessor is inspired by PDEX (now the desktop's main Java mode), which uses Eclipse's JDT DOM library for preprocessing. Among other things, this lets us pick up all of the syntax errors at once, and even proceed to detect compilation errors by fixing the simple syntax problems on the fly.

Another significant problem with the old preprocessor was that it did not store the mappings of original code to preprocessed code and vice versa. This means that highlighting errors found in preprocessed code back in the original code was a non-trivial task. Again inspired by PDEX, we now have a text transform mapper (build/TextTransform.java) for keeping track of changes and mapping ranges between the two sets of code. This change will help not only with the preprocessor, but also with mapping compiler problems back to the original code. PDEX has its own text transform implementation, but it was found to be too difficult to integrate this into APDE, so I just wrote my own. There may still be problems in my implementation, but these problems remain to be seen.

These new JDT features required a lot of additional JARs. That is not a good thing. Hopefully ProGuard does a good job of treeshaking. Another point of interest is that we have passed the dex method limit, so multidex is now enabled. Will have to test multidex extensively to make sure that it doesn't break anything (especially on 4.x) because we do some sketchy stuff with reflection.

In its present state, the preprocessor is mostly functional, but I am currently experiencing difficulties picking up the correct error ranges for preprocessor problems. This should be fairly close to integration into the full build process, but there is still a lot of testing to do before it is ready for prime time.

A random note: I was not able to use the latest version of the Java mode JAR for reasons that I still do not understand. I think there was a problem with Java 8 compatibility or something like that. This means that APDE doesn't have access to PDEX. This is OK, though, because most of the functionality is being completely rewritten anyway.