Work toward new modular, parallelized, and incremental build system.
The build system is contained in the package com.calsignlabs.apde.build.dag. When completed it should allow for faster builds than before and support real-time compilation to detect errors as the user types.

There are still a lot of problems with the current implementation, but it is functional. It supports preview build only, i.e. it doesn't support apps, wallpapers, etc., and it does not support contributed libraries, among other things.