Move undo and redo to the tab bar so that they are always visible.
Undo and redo were previously buried in the action overflow menu on all devices except for those with large screens (e.g. landscape tablets). Undo/redo is a fairly important feature to have easily accessible, so it was deemed necessary to make it available at all times with one tap (instead of two). The best position for these two buttons is to the right of the tab bar, under the action items in the toolbar. This takes space away from the tabs, but the benefits outweigh the costs here.

Making this change required minorly re-jigging the ViewPager, which in turn required adjusting the way that code/console area heights are calculated. I think I have ironed out all of the kinks, but there may still be issues because these calculations are notoriously buggy. In particular, these changes should be tested in emulators.