Implement timeout-based auto-compile. Probably going to stash modular build.
The modular build is close to working but there are a lot of problems with it.
Instead the old build system can now stop after compiling (ECJ) to report
errors. When auto-compilation is enabled, APDE automatically compiles the code
in real-time and highlights problems as you type.

There is some concern about how well this will work on slower devices; testing
is needed.