Upgrade sketch properties to fragment-based, androidx preferences.
The sketch properties activity was a bit of a dinosaur, but now it's shiny and
modern. We have split into a separate SketchPropertiesActivity and
SketchPropertiesFragment, although they are still tightly coupled.

Also got rid of the sketch management controls (new, rename, delete, copy, move
to sketchbook) that were present in sketch properties because they were
buggy and redundant (with the same controls in the editor activity).

Still need to test this on older platforms in the emulator.