Seriously revamp file selection.
aFileChooser is gone. It was old and deprecated, didn't really work, and mostly
just delegated to the storage access framework (SAF) anyway.

The main problem with the way that we selected files in the past was that we
expected to be able to get an absolute path to the file in all circumstances,
but this isn't how Android is meant to operate. Instead of the file scheme,
Android prefers the content scheme (i.e. Uris beginning with "content://").
Most file managers and other applications serve up data this way, which was
breaking APDE whenever it tried to select files because it is difficult to
convert a content Uri into a path (which is by design, for security reasons).
The Android APIs allow us to convert a content Uri into an input stream, so
it was necessary to rework all of the file selection pieces to work in terms
of streams instead of files.

We are now calling into the SAF directly. On top of making all file selection
"just" work, we also have a reasonable interface for creating new files through
the SAF. This will need to be tested extensively on older devices though.

As a side-effect of these changes, APDE now supports selecting files from other
applications, like Google Drive, seamlessly. It's actually kind of cool.

The affected parts of APDE are as follows: adding files (sketch properties),
changing the sketch icon (sketch properties), importing libraries (library
manager), DX dexer (library manager), and keystore creation/selection (export
signed package).

It may be a good idea to create a simple file manager which fully supports
the SAF so that users without another file manager can still select files and,
more importantly, create new files. For some reason, none of the file managers
that I have tested with seem to support creating files through the SAF (and I
have tested with like fifteen file managers), although Google Drive does.