Add instrumented test suite.
Add six instrumented test classes using Espresso and UI Automator. Tested in
emulators for Android 4.4 to 11.0 and it is mostly reliable, but still a work
in progress. Need to expand with more tests, get it to work on Android 12.0,
fix some issues that have been uncovered by the tests, write an additional
unit test suite, and run the tests through CI on GitHub Actions.