Add support for AAPT2.
Motivated primarily by my inability to run an AAPT process on Android
12 due to new security measures that I don't fully understand, I have
instead packaged AAPT2 for APDE. AAPT2 has the advantage of built-in
JNI bindings, so a subprocess invocation is not necessary and it can
be used on Android 12.

It may have been possible to add JNI bindings to AAPT instead, but
AAPT2 is newer and better and we will probably need it in the future
to support things like building app bundles.

I have decided to be more disciplined than I was in the past and am
now including the AOSP patch and a script which can be used to build
AAPT2 from the AOSP source. These files are located in aapt2_build and
have their own README. The built shared libraries and binaries are in
APDE/src/main/jniLibs.

Unfortunately, I haven't gotten the shared library to work on all
Android versions, but it definitely does work for Android 9+. Thus I
am also packaging a binary version of AAPT2 for older versions of
Android. This makes the APK size larger, but it's the best compromise
I could come up with for now. The binary version is named
libaapt2_bin.so even though it is not a shared library because that
was necessary to get Android to automatically copy it out of jniLibs
on install.

Aapt2Jni contains the JNI bindings for the shared library. It includes
support for getting the errors and warnings from AAPT2, which is
exciting because we didn't previously have access to them through the
binary. Aapt2Wrapper wraps Aapt2Jni or the binary depending on the
Android version.

I am still working on migrating the various build systems to use
AAPT2, so more changes are forthcoming.